generalise asset management within AssetManagerPLugin
there are many types of entity in the game that should read at least some of their configuration from an external asset. currently this includes birds, but may extend to other entities such players, food, power-ups etc.
this generalises the systems required to denote an asset as being related to an entity, to load it from a `.ron` (or other built in type) on the file system, and to allow entity specific subsystems to hook into this load/reload cycle with minimum fuss.