Fix birds not spawning
There was a race condition present on the asset loaded event. One handler which should run first cleared the old level from the scene. Another added stuff to the 'new' scene. Issue arose when the latter ran before the former. I previously added `LevelEvent` which is fired when the new level is ready to be populated to handle this case, so have used that.